Add mysql-server recipe.
For the most part, this is copied from the Chef recipe but it seems to
work perfectly. The configuration is fairly memory-hungry (default
key_buffer is 16M, but the Chef recipe has it set to 256M), so I'm
guessing this is intended for small dedicated nodes, or standalone nodes
with > 1GB RAM. More research needed.